# openapi.yaml
openapi: 3.0.3
info:
  title: School Management System API
  description: API for managing users, documents, and tasks in a school management system
  version: 1.0.0
  contact:
    name: API Support
    email: support@schoolmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /tasks:
    get:
      summary: Get tasks
      description: Retrieve tasks with optional search and filter
      tags:
        - Tasks
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: assignee
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      description: Create a new task
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid task data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Subtask:
      type: object
      properties:
        id:
          type: string
          example: "subtask-1-1"
        title:
          type: string
          example: "Design database schema"
        description:
          type: string
          example: "Create ER diagram and database schema for student information"
        status:
          type: string
          enum: [todo, in_progress, done]
          example: "done"
        assignee:
          type: string
          example: "John Smith"
        assigneeId:
          type: integer
          example: 1
        estimatedHours:
          type: integer
          example: 8
        actualHours:
          type: integer
          example: 10
        startedDate:
          type: string
          format: date
          example: "2024-09-19"
        completedDate:
          type: string
          format: date
          example: "2024-09-18"
      required:
        - title
        - description
        - status
        - assignee
        - assigneeId
        - estimatedHours
    Task:
      type: object
      properties:
        id:
          type: string
          example: "task-1"
        title:
          type: string
          example: "Student Registration System"
        description:
          type: string
          example: "Develop a comprehensive student registration system for the new academic year"
        status:
          type: string
          enum: [todo, in_progress, done]
          example: "todo"
        priority:
          type: string
          enum: [low, medium, high]
          example: "high"
        assignee:
          type: string
          example: "John Smith"
        assigneeId:
          type: integer
          example: 1
        reporter:
          type: string
          example: "Sarah Wilson"
        reporterId:
          type: integer
          example: 4
        createdDate:
          type: string
          format: date
          example: "2024-09-15"
        dueDate:
          type: string
          format: date
          example: "2024-10-15"
        completedDate:
          type: string
          format: date
          example: "2024-09-18"
        estimatedHours:
          type: integer
          example: 40
        actualHours:
          type: integer
          example: 30
        labels:
          type: array
          items:
            type: string
          example: ["backend", "database", "urgent"]
        subtasks:
          type: array
          items:
            $ref: '#/components/schemas/Subtask'
      required:
        - title
        - description
        - status
        - priority
        - assignee
        - assigneeId
        - reporter
        - reporterId
        - createdDate
        - dueDate
        - estimatedHours
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "An error occurred"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []